pyproject.toml
[project]
name = "bicker-bot"
version = "0.1.0"
description = "Sibling AI bots (Merry & Hachiman) that bicker on IRC"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.40.0",
    "google-genai>=1.0.0",
    "chromadb>=0.5.0",
    "sentence-transformers>=3.0.0",
    "pydle>=1.0.0",
    "pyyaml>=6.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "hypothesis>=6.100.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.8.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
bicker-bot = "bicker_bot.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/bicker_bot"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
